feat: better build scripts

+++ scripts/clean.sh
@@ -1,12 +1,26 @@
 #!/bin/sh
 
-rm -f CMakeCache.txt Makefile cmake_install.cmake
-rm -rf CMakeFiles
-rm -rf build/CMakeFiles
-for item in "build"/*
-do
-  [ -d "$item" ] && rm -rf $item/build_dir/* && rm -rf $item/build_dir/.*
-done
-rm -f build/cmake_install.cmake build/Makefile
+ROOT_BUILD_DIR=build
 
-rm -rf target/
+rm -rf $ROOT_BUILD_DIR/CMakeFiles
+
+TARGETS="$@"
+
+clear_folder() {
+  local folder="$1"
+  [ -d "$folder" ] && rm -rf "$folder/build_dir/"* && rm -rf "$folder/build_dir/".*
+}
+
+if [ -z "$TARGETS" ]; then
+  rm -f CMakeCache.txt Makefile cmake_install.cmake
+  rm -rf CMakeFiles
+  for item in "$ROOT_BUILD_DIR"/*; do
+    clear_folder "$item"
+  done
+  rm -f $ROOT_BUILD_DIR/cmake_install.cmake $ROOT_BUILD_DIR/Makefile
+  rm -rf target/
+else
+  for target in $TARGETS; do
+    clear_folder "$ROOT_BUILD_DIR/$target"
+  done
+fi;
